+++ src/constants/uiMessages.js
@@ -1,4 +1,4 @@
-export const SCRIPT = {
+export const UI_MESSAGES = {
   NAME_STRING_INPUT: '경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)\n',
   NUMBER_OF_ATTEMPT_INPUT: '시도할 횟수는 몇 회인가요?\n',
   WINNER_IS: '최종 우승자 : ',
