+++ src/models/index.js
@@ -1,0 +1,3 @@
+export { default as RacingGame } from './RacingGame.js';
+export { default as Cars } from './Cars.js';
+export { default as Car } from './Car.js';
