+++ src/utils/index.js
@@ -1,7 +1,3 @@
-export { default as RacingGame } from './RacingGame.js';
-export { default as Cars } from './Cars.js';
-export { default as Car } from './Car.js';
-
 export { default as UserInterface } from './UserInterface.js';
 export { default as Validator } from './Validator.js';
 export { default as GameUtils } from './GameUtils.js';
